requirements.txt
apache-airflow-providers-amazon==5.0.0
boto3==1.26.0
pandas==1.3.5
sqlalchemy==1.4.0
psycopg2-binary==2.9.3